agent_core/requirements.txt
livekit-agents>=1.2.0
google-generativeai>=0.8.0
google-cloud-texttospeech>=2.16.0
mem0ai>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.10.0
